pyproject.toml
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.black]
line-length = 110
target-version = ["py38"]

[tool.isort]
line_length = 110
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
known_first_party = "" # top packages
known_local_folder = "tests,tapioca_box_delivery"
sections = "FUTURE,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"
default_section = "THIRDPARTY"

[tool.poetry]
name = "tapioca_box_delivery"
version = "1.0.0"
description = "tapioca box delivery client"
authors = ["Developers <wilsontads@gmail.com>"]
repository = "tapioca-box-delivery"
keywords = [ "yaeeh", "tapioca" ]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Topic :: Software Development :: Libraries",
]
exclude = [ "docs", "tests*"]
include = ["*.rst", "*.txt", "*.py", "*.png", "*.dot", "Makefile"]

[[tool.poetry.source]]
name = "fury"
url = "https://1rSTPY-qDFbggOUEkyAQzZSRpR9Hzt1R7Q:@pypi.fury.io/wilsonmoraes/"

[tool.poetry.dependencies]
python = "^3.9"
tapioca-wrapper = ">=2.0.1"


[tool.poetry.dev-dependencies]
coverage = "*"
pre-commit = "*"
pytest = "*"
pytest-cov = "*"
pytest-randomly = "*"
pytest-deadfixtures = "*"
ipython = "*"
safety = "*"
docproduct="*"